move file parsing part into the parser/file package + run file parser specific tests
